Animation on initial rendering fix
Animation started at the load of the page, but now it starts doing it's thing after clicking the button.
The fix is not preety, but i have no idea how to do it better, mayba MAYBE i will be able to make this part of the code better, but i don't think so. Anyway - it works, don't cause issues, so it's good.